requirements.txt
Django==1.9
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
django-vote